+++ gradle/libs.versions.toml
@@ -1,7 +1,7 @@
 [versions]
 grpcKotlinMultiplatform = "0.3.0"
 
-androidGradlePlugin = "8.1.0"
+androidGradlePlugin = "8.0.2"
 androidCompileSdk = "31"
 androidMinSdk = "21"
