+++ grpc-mp-test/src/jsTest/kotlin/io/github/timortel/kotlin_multiplatform_grpc_plugin/test/JSSerializationTest.kt
@@ -1,33 +1,28 @@
 package io.github.timortel.kotlin_multiplatform_grpc_plugin.test
 
-import io.github.timortel.kotlin_multiplatform_grpc_lib.message.JSImpl
 import io.github.timortel.kotlin_multiplatform_grpc_lib.message.KMMessage
 import io.github.timortel.kotlin_multiplatform_grpc_lib.message.MessageDeserializer
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.test.basic_messages.*
-import kotlin.test.Test
-import kotlin.test.assertEquals
 
 class JSSerializationTest : SerializationTest() {
 
-    private fun <T : JSImpl, J : KMMessage> serializeImpl(msg: T, deserializer: MessageDeserializer<T>, creator: (T) -> J): J {
-        return creator(deserializer.deserializeBinary(msg.serializeBinary()))
+    private fun <T : KMMessage> serializeImpl(msg: T, deserializer: MessageDeserializer<T, dynamic>): T {
+        return deserializer.deserialize(msg.serialize())
     }
 
     override fun serialize(message: KMLongMessage): KMLongMessage =
-        serializeImpl(message.jsImpl, JS_LongMessage.Companion, ::KMLongMessage)
+        serializeImpl(message, KMLongMessage.Companion)
 
     override fun serialize(message: KMRepeatedLongMessage): KMRepeatedLongMessage =
-        serializeImpl(message.jsImpl, JS_RepeatedLongMessage.Companion, ::KMRepeatedLongMessage)
+        serializeImpl(message, KMRepeatedLongMessage.Companion)
 
-    override fun serialize(message: KMScalarTypes): KMScalarTypes =
-        serializeImpl(message.jsImpl, JS_ScalarTypes.Companion, ::KMScalarTypes)
+    override fun serialize(message: KMScalarTypes): KMScalarTypes = serializeImpl(message, KMScalarTypes.Companion)
 
     override fun serialize(message: KMMessageWithSubMessage): KMMessageWithSubMessage =
-        serializeImpl(message.jsImpl, JS_MessageWithSubMessage.Companion, ::KMMessageWithSubMessage)
+        serializeImpl(message, KMMessageWithSubMessage.Companion)
 
     override fun serialize(message: KMMessageWithEverything): KMMessageWithEverything =
-        serializeImpl(message.jsImpl, JS_MessageWithEverything.Companion, ::KMMessageWithEverything)
+        serializeImpl(message, KMMessageWithEverything.Companion)
 
-    override fun serialize(message: KMOneOfMessage): KMOneOfMessage =
-        serializeImpl(message.jsImpl, JS_OneOfMessage.Companion, ::KMOneOfMessage)
+    override fun serialize(message: KMOneOfMessage): KMOneOfMessage = serializeImpl(message, KMOneOfMessage.Companion)
 }
