+++ grpc-multiplatform-lib/src/commonMain/kotlin/io/github/timortel/kotlin_multiplatform_grpc_lib/message/KMMessage.kt
@@ -3,9 +3,4 @@ package io.github.timortel.kotlin_multiplatform_grpc_lib.message
 /**
  * Base interface for all grpc messages. All messages are automatically generated by the gradle plugin based on your proto definitions.
  */
-expect interface KMMessage {
-    /**
-     * The size this message takes up in a byte array.
-     */
-    val requiredSize: Int
-}
+expect interface KMMessage
