+++ grpc-multiplatform-lib/src/jsMain/kotlin/io/github/timortel/kotlin_multiplatform_grpc_lib/io/TypeUtil.kt
@@ -1,0 +1,5 @@
+package io.github.timortel.kotlin_multiplatform_grpc_lib.io
+
+import org.khronos.webgl.Uint8Array
+
+fun Uint8Array.toByteArray(): ByteArray = unsafeCast<ByteArray>()
