--- grpc-multiplatform-lib/src/jsMain/kotlin/io/github/timortel/kotlin_multiplatform_grpc_lib/message/JSImpl.kt
@@ -1,5 +1,0 @@
-package io.github.timortel.kotlin_multiplatform_grpc_lib.message
-
-interface JSImpl {
-    fun serializeBinary(): dynamic
-}
