--- grpc-multiplatform-lib/src/jsMain/kotlin/io/github/timortel/kotlin_multiplatform_grpc_lib/message/MessageDeserializer.kt
@@ -1,5 +1,0 @@
-package io.github.timortel.kotlin_multiplatform_grpc_lib.message
-
-interface MessageDeserializer<T : JSImpl> {
-    fun deserializeBinary(bytes: dynamic): T
-}
