--- plugin/src/main/java/io/github/timortel/kotlin_multiplatform_grpc_plugin/generate_mulitplatform_sources/generators/common/JvmCommonFunctionGenerator.kt
@@ -1,14 +1,0 @@
-package io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.common
-
-import com.squareup.kotlinpoet.*
-import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.content.ProtoMessage
-
-class JvmCommonFunctionGenerator(private val builder: FileSpec.Builder) : CommonFunctionGenerator() {
-
-    override fun addFunction(funSpec: FunSpec) {
-        builder.addFunction(funSpec)
-    }
-
-    override fun getNativePlatformType(message: ProtoMessage): TypeName = message.jvmType
-
-}
