+++ plugin/src/main/java/io/github/timortel/kotlin_multiplatform_grpc_plugin/generate_mulitplatform_sources/generators/proto3_ios_file_writer.kt
@@ -2,11 +2,11 @@ package io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatfo
 
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.content.ProtoFile
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.dsl.IosJvmDslBuilder
-import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.proto_file.IOSProtoFileWriter
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.proto_file.IosProtoFileWriter
 import java.io.File
 
 fun writeIOSFiles(protoFile: ProtoFile, iosOutputDir: File) {
-    IOSProtoFileWriter(protoFile).writeFile(iosOutputDir)
+    IosProtoFileWriter(protoFile).writeFile(iosOutputDir)
 
     //JVM helper
 //    FileSpec
