+++ plugin/src/main/java/io/github/timortel/kotlin_multiplatform_grpc_plugin/generate_mulitplatform_sources/generators/proto3_ios_service_writer.kt
@@ -2,9 +2,9 @@ package io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatfo
 
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.content.ProtoFile
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.content.ProtoService
-import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.service.IOSServiceWriter
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.service.IosServiceWriter
 import java.io.File
 
 fun writeIOSServiceFile(protoFile: ProtoFile, service: ProtoService, iosOutputFolder: File) {
-    IOSServiceWriter.writeServiceStub(protoFile, service, iosOutputFolder)
+    IosServiceWriter.writeServiceStub(protoFile, service, iosOutputFolder)
 }
