+++ plugin/src/main/java/io/github/timortel/kotlin_multiplatform_grpc_plugin/generate_mulitplatform_sources/generators/proto_file/JsProtoFileWriter.kt
@@ -1,44 +1,83 @@
 package io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.proto_file
 
-import com.squareup.kotlinpoet.*
-import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.content.*
+import com.squareup.kotlinpoet.CodeBlock
+import com.squareup.kotlinpoet.Dynamic
+import com.squareup.kotlinpoet.FunSpec
+import com.squareup.kotlinpoet.TypeName
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.CodedInputStream
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.CodedOutputStream
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.JSPB_BINARY_READER
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.JSPB_BINARY_WRITER
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.content.MapType
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.content.ProtoFile
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.content.ProtoMessage
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.content.ProtoMessageAttribute
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.Const
-import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.map.JsMapMessageMethodGenerator
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.map.IosJvmMapMessageMethodGenerator
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.map.MapMessageMethodGenerator
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.oneof.IosJvmOneOfMethodAndClassGenerator
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.oneof.JsOneOfMethodAndClassGenerator
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.oneof.OneOfMethodAndClassGenerator
-import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.repeated.JsRepeatedMessageMethodGenerator
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.repeated.IosJvmRepeatedMessageMethodGenerator
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.repeated.RepeatedMessageMethodGenerator
-import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.scalar.JsScalarMessageMethodGenerator
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.scalar.IosJvmScalarMessageMethodGenerator
 import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.generators.scalar.ScalarMessageMethodGenerator
+import io.github.timortel.kotlin_multiplatform_grpc_plugin.generate_mulitplatform_sources.writeMap
 
-class JsProtoFileWriter(private val protoFile: ProtoFile) : ProtoFileWriter(protoFile, true), DefaultChildClassName {
-    override val scalarMessageMethodGenerator: ScalarMessageMethodGenerator = JsScalarMessageMethodGenerator
+class JsProtoFileWriter(protoFile: ProtoFile) : ActualProtoFileWriter(protoFile) {
 
-    override val repeatedMessageMethodGenerator: RepeatedMessageMethodGenerator = JsRepeatedMessageMethodGenerator
+    override val scalarMessageMethodGenerator: ScalarMessageMethodGenerator
+        get() = IosJvmScalarMessageMethodGenerator
+    override val repeatedMessageMethodGenerator: RepeatedMessageMethodGenerator
+        get() = IosJvmRepeatedMessageMethodGenerator
+    override val oneOfMethodAndClassGenerator: OneOfMethodAndClassGenerator
+        get() = JsOneOfMethodAndClassGenerator
+    override val mapMessageMethodGenerator: MapMessageMethodGenerator
+        get() = IosJvmMapMessageMethodGenerator
 
-    override val oneOfMethodAndClassGenerator: OneOfMethodAndClassGenerator = JsOneOfMethodAndClassGenerator
+    override val serializeFunctionCode: CodeBlock
+        get() = CodeBlock.builder()
+            .addStatement("val writer = %T()", JSPB_BINARY_WRITER)
+            .addStatement("val stream = %T(writer)", CodedOutputStream)
+            .addStatement("serialize(stream)")
+            .addStatement("return writer.getResultBuffer()")
+            .build()
 
-    override val mapMessageMethodGenerator: MapMessageMethodGenerator = JsMapMessageMethodGenerator
+    override val serializedDataType: TypeName = Dynamic
 
-    override fun applyToClass(builder: TypeSpec.Builder, message: ProtoMessage, messageClassName: ClassName) {
-        val paramName = Const.Message.Constructor.JS.PARAM_IMPL
+    override val deserializeFunctionCode: CodeBlock
+        get() = CodeBlock
+            .builder()
+            .addStatement(
+                "val stream = %T(%T(data))",
+                CodedInputStream,
+                JSPB_BINARY_READER
+            )
+            .addStatement(
+                "return %N(stream)",
+                Const.Message.Companion.WrapperDeserializationFunction.NAME
+            )
+            .build()
 
-        builder.addProperty(
-            PropertySpec
-                .builder(paramName, message.jsType, KModifier.PUBLIC, KModifier.OVERRIDE)
-                .initializer(paramName)
-                .build()
-        )
+    override fun buildMapAttributeSerializeCode(
+        builder: FunSpec.Builder,
+        message: ProtoMessage,
+        attr: ProtoMessageAttribute,
+        mapType: MapType
+    ) {
+        builder.apply {
+            addCode(
+                "%M(%N, %L, %N, ",
+                writeMap,
+                Const.Message.IOS.SerializeFunction.STREAM_PARAM,
+                attr.protoId,
+                Const.Message.Attribute.propertyName(message, attr)
+            )
 
-        builder.primaryConstructor(
-            FunSpec
-                .constructorBuilder()
-                .addParameter(paramName, message.jsType)
-                .build()
-        )
+            addMapKeyTypeSerializationCode(mapType.keyTypes)
+            addCode(", ")
+            addMapValueTypeSerializationCode(mapType.valueTypes)
+            addCode(")\n")
+        }
     }
-
-    override fun getChildClassName(parentClass: ClassName?, childName: String): ClassName =
-        getChildClassName(parentClass, childName, protoFile.pkg)
 }
